% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{compute_partitions}
\alias{compute_partitions}
\title{C partitions generator}
\usage{
compute_partitions(v, K = 2)
}
\arguments{
\item{v}{A feature vector ( a column of V)}
}
\value{
An array that represents a cube. Each slice in the cube is a matrix C.
}
\details{
Compute all possible matrices C that explain a feature vector
}
