% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{compute_partitions_N}
\alias{compute_partitions_N}
\title{title: Compute C partitions for the equation
where every term has all N}
\usage{
compute_partitions_N(V, K = 2)
}
\description{
title: Compute C partitions for the equation
where every term has all N
}
