% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{dnegmul_prints}
\alias{dnegmul_prints}
\title{Negative Multinomial distribution with weight parameters}
\usage{
dnegmul_prints(x, w = 1:length(x), alpha = 1, beta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of counts}

\item{w}{vector of weights in every dimension}

\item{alpha}{number of failure experiments}

\item{beta}{weight of the failure dimension.}
}
