% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL-NMF_Lee_Seung.R
\name{KL_divergence_Lee}
\alias{KL_divergence_Lee}
\title{Compute the KL divergence}
\usage{
KL_divergence_Lee(A, B)
}
\arguments{
\item{A}{first matrix}

\item{B}{second matrix}
}
\details{
This function is useful for debugging
}
