% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{likelihood_C_W}
\alias{likelihood_C_W}
\title{Product of independent Negative Multinomial Distributions}
\usage{
likelihood_C_W(C, W, alpha = 1, beta = 1, log = FALSE)
}
\arguments{
\item{C}{matrix of counts}

\item{W}{matrix of weights in every dimension}

\item{alpha}{number of failure experiments}

\item{beta}{1/weight of the failure dimension.}
}
\details{
Column k of C is associated to column k of W.
}
