% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{likelihood_C_W_manual}
\alias{likelihood_C_W_manual}
\title{manual computation for F=1, K=2, N=2, and e.g. v=[1,6]}
\usage{
likelihood_C_W_manual(C, w, alpha = 10)
}
\arguments{
\item{C:}{c vectors to be used. C is N x K dimensions}
}
\details{
this is the most simple example that shows two modes
The goal of this function is to "debug" the model and understand 
at what point the bimodality emerges.
}
