% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeform.R
\name{likelihood_V_W}
\alias{likelihood_V_W}
\title{likelihood p(V | W)}
\usage{
likelihood_V_W(V, W, alpha, beta, log = TRUE, vectorizable = FALSE)
}
\arguments{
\item{V}{a full matrix with one observation per column}
}
\value{
log probability
}
\description{
computes p(V|W)
}
\details{
Computes the likelihood of a full V. Since each column
is independent, the likelihood of the whole matrix V is the product of the
likelihoods of each column in V.
}
