% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_image_file}
\alias{load_image_file}
\title{Loads the data matrix (V matrix) from the training file}
\usage{
load_image_file(filename)
}
\value{
A matrix with one row per image
}
\description{
Loads the data matrix (V matrix) from the training file
}
