% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_synthetic}
\alias{load_synthetic}
\title{Generates some synthetic matrix (ex: user x movie)
From a chosen W (ex: user x topic)}
\usage{
load_synthetic()
}
\value{
A matrix with a clear structure (ex: block diagonal matrix)
}
\description{
Generates some synthetic matrix (ex: user x movie)
From a chosen W (ex: user x topic)
}
