% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{marginal_likelihood_basic}
\alias{marginal_likelihood_basic}
\title{Marginal likelihood with samples from the prior}
\usage{
marginal_likelihood_basic(V, W, alpha = 1, beta = 1, nsamples = 100)
}
\arguments{
\item{V}{data matrix}

\item{W}{dictionary matrix}

\item{alpha}{parameter for the prior}

\item{beta}{parameter for the prior}

\item{nsamples}{number of samples from the Importance distribution}
}
