% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{marginal_likelihood_Chib}
\alias{marginal_likelihood_Chib}
\title{Likelihood p(V | W) with Chib's method}
\usage{
marginal_likelihood_Chib(V, W, H, C, alpha, beta)
}
\arguments{
\item{V}{original matrix}

\item{W}{estimated dictionary matrix}

\item{H}{any H matrix}

\item{C}{samples of matrix of auxiliary variables c}
}
\details{
Computes equation 22 of the paper Dikmen, Févotte (2012)
to compute p(H | V, W) we use the samples from the last iterations
since these are closer to the true posterior than the older ones where 
W was not optimal yet.
}
