% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL-NMF_Lee_Seung.R
\name{nmf_Lee}
\alias{nmf_Lee}
\title{NMF with Lee and Saung multiplicative updates}
\usage{
nmf_Lee(V, K, W, H, maxiters = 100)
}
\arguments{
\item{V}{matrix to factorize}

\item{K}{number of latent factors (or dimensions)}

\item{W}{initial W matrix}

\item{H}{initial H matrix}

\item{maxiters}{maximum number of iterations}
}
\details{
The number of iterations is set to maxiters.
In the future, there should be a convergence check in case
the KL converges before maxiters
}
