% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMLE_MCEM.R
\name{nmf_mmle_mcem}
\alias{nmf_mmle_mcem}
\title{NMF with MMLE}
\usage{
nmf_mmle_mcem(V, initW, maxiters = 100, nsamples = 50, algo = "gibbs",
  mincrease = "constant", alpha = 1, beta = 1, mstep = "C", L = 10,
  epsilon = 0.5, burnin = 0.5)
}
\arguments{
\item{V}{matrix to be factorized}

\item{maxiters}{maximum number of iterations}

\item{nsamples}{number of samples per iteration in the E-step}

\item{K}{number of latent factors (or dimensions)}

\item{W}{initial matrix W}

\item{mc}{mcem or saem.}
}
\value{
W last estimator of W
        traces$H: a K x N x samples tensor
        traces$W: a K x samples - matrix vector with the evolution of row norms
}
\details{
The number of iterations is set to maxiters.
In the future, there should be a convergence check in case
the KL converges before maxiters.

If saem is used, then the number of samples gradually increases until reaching
nsamples for the last iterations
}
