% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{plot_image}
\alias{plot_image}
\title{Plot an image from a given row in the data matrix}
\usage{
plot_image(data, n)
}
\arguments{
\item{n}{the number of the column to plot}

\item{V}{the whole data matrix}
}
\description{
Plot an image from a given row in the data matrix
}
\details{
It gets the n-th column, loads it to a 28x28 matrix and then plots it
}
