% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multiple_W}
\alias{plot_multiple_W}
\title{Plot four matrices}
\usage{
plot_multiple_W(..., cols = 4)
}
\arguments{
\item{a}{set of matrices}
}
\description{
Plot four matrices in the same row using ggplot
}
\details{
This method is useful to compare, for instance, the different W
given by different choices of $K$ (number of latent dimensions) in NMF
}
