% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_H.R
\name{posterior_h}
\alias{posterior_h}
\title{posterior of a column of H}
\usage{
posterior_h(h, v, W, alpha = 1, beta = 1, log = FALSE)
}
\arguments{
\item{h}{a column vector from H, of length K}
}
\description{
proportionality constants are ignored. Thus, the posterior is
unnormalized.
}
\details{
For the sake of simplicity, this is done by multiplying the gamma prior
and the poisson likelihood
}
