% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gibbs.R
\name{sample_gibbs}
\alias{sample_gibbs}
\title{Gibbs sampler of h_n using auxiliary variables C}
\usage{
sample_gibbs(v_n, W, h_n, alpha = 1, beta = 1, iter = 100)
}
\arguments{
\item{v_n}{n-th column of matrix V (F x 1)}

\item{W}{dictionary matrix (F x K)}

\item{h_n}{(K x 1) initial value h_n (n-th column of matrix H)}

\item{alpha}{parameter alpha}

\item{beta}{parameter beta}

\item{iter}{number of samples (or "iterations")}
}
