% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gibbs.R
\name{sample_gibbs_h}
\alias{sample_gibbs_h}
\title{Conditional sampler of h_kn}
\usage{
sample_gibbs_h(w_k, c_kn, alpha = 1, beta = 1)
}
\arguments{
\item{w_k}{k-th column of matrix W}

\item{c_kn}{vector in position (k,n) of tensor C [F x K x N]}
}
