% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gibbs_z.R
\name{sample_gibbs_z}
\alias{sample_gibbs_z}
\title{Gibbs sampler of h_n using auxiliary variables C}
\usage{
sample_gibbs_z(v_n, W, C_n, alpha = 1, beta = 1, iter = 100)
}
\arguments{
\item{v_n}{n-th column of matrix V (F x 1)}

\item{W}{dictionary matrix (F x K)}

\item{C_n}{(F x K) initial value C_n}

\item{alpha}{parameter alpha}

\item{beta}{parameter beta}

\item{iter}{number of samples (or "iterations")}
}
