% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hamiltonian.R
\name{sample_hamiltonian_h}
\alias{sample_hamiltonian_h}
\title{Hamiltonian sample of a column h_n}
\usage{
sample_hamiltonian_h(v_n, W, h_n_current, epsilon = 0.1, L = 10,
  iter = 100, alpha, beta)
}
\arguments{
\item{v_n}{n-th column of matrix V (F x 1)}

\item{W}{dictionary matrix (F x K)}

\item{h_n_current}{(K x 1) initial value h_n (n-th column of matrix H)}

\item{epsilon}{length of each Hamiltonian step}

\item{L}{number of Hamiltonian steps at each iteration ("path length")}

\item{iter}{number of samples (or "iterations")}
}
