% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hamiltonian_old.R
\name{sample_hamiltonian_h_old}
\alias{sample_hamiltonian_h_old}
\title{Hamiltonian sample of a column h_n}
\usage{
sample_hamiltonian_h_old(v_n, W, current_q, epsilon = 0.1, L = 10,
  iter = 100)
}
