% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metropolis.R
\name{sample_metropolis_h_reparam}
\alias{sample_metropolis_h_reparam}
\title{sample h_{n} given all the other h in the same column 
with Metroplis Hastings. Use a log-transform reparamatrization}
\usage{
sample_metropolis_h_reparam(v_n, W, h_n_current, alpha = 1, beta = 1,
  step = 1, iter = 100)
}
\arguments{
\item{v_n}{n-th column of matrix V (F x 1)}

\item{W}{dictionary matrix (F x K)}

\item{h_n_current}{(K x 1) initial value h_n (n-th column of matrix H)}

\item{alpha}{parameter alpha}

\item{beta}{parameter beta}

\item{step}{step size for the proposal distribution}

\item{iter}{number of samples (or "iterations")}
}
