% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL-NMF_Lee_Seung.R
\name{update_h_Lee}
\alias{update_h_Lee}
\title{Update H_kn}
\usage{
update_h_Lee(k, n, V, W, H)
}
\arguments{
\item{k}{row of H}

\item{n}{column of H}

\item{V}{matrix being factorized}

\item{W}{dictionary matrix}

\item{H}{activation coefficients matrix}
}
\details{
Updates the given position using Lee and Seung multiplicative updates
}
