% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL-NMF_Lee_Seung.R
\name{update_w_Lee}
\alias{update_w_Lee}
\title{Update W_fk}
\usage{
update_w_Lee(f, k, V, W, H)
}
\arguments{
\item{f}{row of W}

\item{k}{column of W}

\item{V}{matrix being factorized}

\item{W}{dictionary matrix}

\item{H}{activation coefficients matrix}
}
\details{
Updates the given position using Lee and Seung multiplicative updates
}
