% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_W_mc_C}
\alias{update_W_mc_C}
\title{Maximize w.r.t W given C}
\usage{
update_W_mc_C(C_samples_mean, alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{parameter alpha}

\item{beta}{parameter beta}

\item{W}{the (whole) current dictionary matrix.}

\item{C_samples}{samples of the C tensor}
}
