% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_W_mc_CH}
\alias{update_W_mc_CH}
\title{Maximize w.r.t W using C,H}
\usage{
update_W_mc_CH(C_samples_mean, H_samples_mean)
}
\arguments{
\item{C_samples}{samples of C tensor (F,K,N)}

\item{H_samples}{samples of H matrix (K,N)}
}
