% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_W_mc_CH_SAEM}
\alias{update_W_mc_CH_SAEM}
\title{Maximize w.r.t W using C,H}
\usage{
update_W_mc_CH_SAEM(C_weighted_sum, H_weighted_sum)
}
\arguments{
\item{C_mean}{weighted means of C (F,K, niter)}

\item{H_samples}{samples of H matrix (K, niter)}
}
