% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_W_mc_H}
\alias{update_W_mc_H}
\title{Maximize w.r.t W using H with multiplicative updates}
\usage{
update_W_mc_H(W, V, H_samples, updates = 1)
}
\arguments{
\item{W}{the current dictionary matrix.}

\item{V}{matrix to be factorized.}

\item{H}{activation coefficients matrix (J,K,N)}
}
