% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_W_mc_H_depprecated}
\alias{update_W_mc_H_depprecated}
\title{Maximize w.r.t W using H with multiplicative updates}
\usage{
update_W_mc_H_depprecated(W, V, H_samples)
}
\arguments{
\item{W}{the (whole) current dictionary matrix.}

\item{V}{matrix to be factorized.}

\item{H}{activation coefficients matrix (J,K,N)}
}
